+++ src/main/scala/stellar/sdk/Network.scala
@@ -410,7 +410,7 @@ trait Network extends LazyLogging {
   }
 
   /**
-    * Fetch a stream of transactions affecting a given account
+    * Fetch a stream of transactions affecting a given account.
     * @param pubKey the relevant account
     * @param cursor optional record id to start results from (defaults to `0`)
     * @param order  optional order to sort results by (defaults to `Asc`)
@@ -432,7 +432,7 @@ trait Network extends LazyLogging {
   }
 
   /**
-    * Fetch a stream of transactions for a given ledger
+    * Fetch a stream of transactions for a given ledger.
     * @param cursor optional record id to start results from (defaults to `0`)
     * @param order  optional order to sort results by (defaults to `Asc`)
     * @see [[https://www.stellar.org/developers/horizon/reference/endpoints/transactions-for-ledger.html endpoint doc]]
@@ -457,6 +457,29 @@ trait Network extends LazyLogging {
     */
   def feeStats()(implicit ex: ExecutionContext): Future[FeeStatsResponse] =
     horizon.get[FeeStatsResponse]("/fee_stats")
+
+  /**
+    * Fetch a stream of payment paths that realise a payment of the requested destination amount, from the specified
+    * account.
+    * @param from the account that wishes to make the payment
+    * @param to the recipient account
+    * @param amount the desired payment amount
+    * @see [[https://www.stellar.org/developers/horizon/reference/endpoints/path-finding.html endpoint doc]]
+    */
+  def paths(from: PublicKeyOps, to: PublicKeyOps, amount: Amount)(implicit ex: ExecutionContext): Future[Seq[PaymentPath]] = {
+    val queryParams = Map(
+      "source_account" -> from.accountId,
+      "destination_account" -> to.accountId
+    ) ++ amountParams("destination", amount)
+    horizon.getSeq[PaymentPath]("/paths", PaymentPathDeserializer, queryParams)
+  }
+
+  private def amountParams(prefix: String, amount: Amount): Map[String, String] = {
+    (amount match {
+      case issuedAmount: IssuedAmount => assetParams(prefix, issuedAmount.asset)
+      case _ => Map.empty[String, String]
+    }) ++ Map(s"${prefix}_amount" -> s"${amount.units}")
+  }
 
   private def assetParams(prefix: String, asset: Asset): Map[String, String] = {
     asset match {