+++ src/main/scala/stellar/sdk/inet/HorizonAccess.scala
@@ -48,6 +48,9 @@ trait HorizonAccess {
   def getStream[T: ClassTag](path: String, de: CustomSerializer[T], cursor: HorizonCursor, order: HorizonOrder, params: Map[String, String] = Map.empty)
                             (implicit ec: ExecutionContext, m: Manifest[T]): Future[Stream[T]]
 
+  def getSeq[T: ClassTag](path: String, de: CustomSerializer[T], params: Map[String, String] = Map.empty)
+                            (implicit ec: ExecutionContext, m: Manifest[T]): Future[Seq[T]]
+
   def getSource[T: ClassTag](path: String, de: CustomSerializer[T], cursor: HorizonCursor, params: Map[String, String] = Map.empty)
                             (implicit ec: ExecutionContext, m: Manifest[T]): Source[T, NotUsed]
 
@@ -125,24 +128,35 @@ class Horizon(call: HttpRequest => Future[HttpResponse])
     else Unmarshal(response.entity).to[JObject].map(HorizonServerError(request.uri, _)).map(Failure(_))
   }
 
+  override def getSeq[T: ClassTag](path: String, de: CustomSerializer[T], params: Map[String, String])
+                                  (implicit ec: ExecutionContext, m: Manifest[T]): Future[Seq[T]] =
+    getStream(path, de, None, None, params)
+
   override def getStream[T: ClassTag](path: String, de: CustomSerializer[T], cursor: HorizonCursor, order: HorizonOrder,
                                       params: Map[String, String] = Map.empty)
+                                     (implicit ec: ExecutionContext, m: Manifest[T]): Future[Stream[T]] =
+    getStream(path, de, Some(cursor), Some(order), params)
+
+  private def getStream[T: ClassTag](path: String, de: CustomSerializer[T], cursor: Option[HorizonCursor],
+                                     order: Option[HorizonOrder], params: Map[String, String])
                                      (implicit ec: ExecutionContext, m: Manifest[T]): Future[Stream[T]] = {
 
     import scala.concurrent.duration._
 
     implicit val formats = DefaultFormats + RawPageDeserializer + de
 
-    val query = Query(params ++ Map(
-      "cursor" -> cursor.paramValue,
-      "order" -> order.paramValue,
-      "limit" -> "100"
-    ))
+    val query = Query(params ++ Seq(
+      cursor.map("cursor" -> _.paramValue),
+      order.map("order" -> _.paramValue),
+      Some("limit" -> "100")
+    ).flatten.toMap)
 
     val requestUri = Uri(path).withQuery(query)
 
-    def next(p: Page[T]): Future[Option[Page[T]]] =
-      (getPage(Uri(p.nextLink)): Future[Page[T]]).map(Some(_))
+    def next(p: Page[T]): Future[Option[Page[T]]] = p.nextLink match {
+      case None => Future.successful(None)
+      case Some(link) => (getPage(Uri(link)): Future[Page[T]]).map(Some(_))
+    }
 
     def stream(ts: Seq[T], maybeNextPage: Future[Option[Page[T]]]): Stream[T] = {
       ts match {
@@ -172,7 +186,7 @@ class Horizon(call: HttpRequest => Future[HttpResponse])
       else call(request)
 
     response.flatMap {
-      case r if r.status == NotFound => Future(Page(Seq.empty[T], uri.toString()))
+      case r if r.status == NotFound => Future(Page(Seq.empty[T], Some(uri.toString())))
       case r                         => Unmarshal(r).to[RawPage].map(_.parse[T])
     }
     .recover { case t: Throwable => throw new RuntimeException(s"Unable to get page for $uri", t) }