+++ src/main/scala/stellar/sdk/inet/Page.scala
@@ -7,16 +7,16 @@ import stellar.sdk.model.response.ResponseParser
 /**
   * A page of results
   */
-case class Page[T](xs: Seq[T], nextLink: String)
+case class Page[T](xs: Seq[T], nextLink: Option[String])
 
-case class RawPage(inner: Seq[JValue], nextLink: String) {
+case class RawPage(inner: Seq[JValue], nextLink: Option[String]) {
   def parse[T](implicit formats: Formats, m: Manifest[T]): Page[T] = Page(inner.map(_.extract[T]), nextLink)
 }
 
 object RawPageDeserializer extends ResponseParser[RawPage]({ o: JObject =>
   implicit val formats = DefaultFormats
 
-  val nextLink = (o \ "_links" \ "next" \ "href").extract[String]
+  val nextLink = (o \ "_links" \ "next" \ "href").extractOpt[String]
   val JArray(records) = o \ "_embedded" \ "records"
 
   RawPage(records, nextLink)